requirement.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
EasyProcess==1.1
entrypoint2==1.1
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
patool==1.12
Pillow==10.0.0
pygame==2.5.1
pyunpack==0.3
requests==2.31.0
soupsieve==2.4.1
tinytag==1.9.0
urllib3==2.0.4
